Removes the vendor director in favour of go modules support
Unfortunately, Go modules has some issues with glfw where it does not
correctly vendor the c source files. Given the stronger version
guarantees with go modules, it was opted to drop the vendor directory
in favour of supporting the latest tooling.

Instructions for building and the circle ci configs have been updated to
match.