Add Windows launcher
The launcher looks at directory "Versions" next to its executable.
It finds the latest version and runs the executable in that directory
with the same name as itself. For instance:

  Aminal.exe <- the launcher
  Versions/
    1.0.0/
      Aminal.exe
    1.0.1/
      Aminal.exe

In this example, running the top-level Aminal.exe (the launcher) starts
Versions/1.0.1/Aminal.exe.

Having a launcher allows Aminal to be updated while it is running. For
example, version 1.0.1 could be downloaded without disturbing running
instances of Aminal 1.0.0.